useDarkMode hook implemented
