+++ src/components/icons/MoonIcon.js
@@ -1,6 +1,6 @@
 export const MoonIcon = () => {
 	return (
-		<svg className='dark:hidden' width='20' height='20' xmlns='http://www.w3.org/2000/svg'>
+		<svg width='20' height='20' xmlns='http://www.w3.org/2000/svg'>
 			<title>Switch to dark mode</title>
 
 			<path
