+++ src/hooks/useDarkMode.js
@@ -1,0 +1,38 @@
+import { useEffect, useState } from 'react';
+
+export const useDarkMode = () => {
+	const darkQuery = window.matchMedia('(prefers-color-scheme: dark)');
+
+	const [theme, setTheme] = useState(localStorage.getItem('theme') ? localStorage.getItem('theme') : darkQuery.matches);
+
+	useEffect(() => {
+		if (theme === 'dark') {
+			document.documentElement.classList.add('dark');
+		} else {
+			document.documentElement.classList.remove('dark');
+		}
+	}, [theme]);
+
+	const handleTheme = () => {
+		console.log('kliknięto');
+		if (theme === 'dark') {
+			setTheme('light');
+			localStorage.setItem('theme', 'light');
+		} else {
+			setTheme('dark');
+			localStorage.setItem('theme', 'dark');
+		}
+	};
+
+	darkQuery.addEventListener('change', e => {
+		if (!('theme' in localStorage)) {
+			if (e.matches) {
+				document.documentElement.classList.add('dark');
+			} else {
+				document.documentElement.classList.remove('dark');
+			}
+		}
+	});
+
+	return [handleTheme, theme];
+};
